Accumulated list of optional arguments is a local variable.
Previously, it was treated as a paremeter to the function, however since
we started following the C calling convetion and operating the stack
like civilized people, it no longer worked. If the list of optional
arguments was empty, that'd mean there were no optional arguments
provided, therefore there was no space on the stack allocated for this
special variable.